requirements_demo.txt
flask>=2.3
flask-socketio>=5.3
ultralytics>=8.0
opencv-python>=4.8
numpy>=1.24,<2
Pillow>=9.5
pyyaml>=6.0
google-generativeai>=0.4
aiohttp>=3.9
